refactor(jobs): split scheduler into focused modules
