+++ src/app/jobs/tasks/trash.rs
@@ -1,7 +1,7 @@
 use super::*;
 use std::{
     fs,
-    path::PathBuf,
+    path::{Path, PathBuf},
     sync::{
         Arc, Condvar, Mutex,
         atomic::{AtomicBool, AtomicU64, Ordering},
